requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.1
sqlparse==0.5.0
